Introduced generic iteration function in result package
A new function, Iter, has been added to the result package. This function is a generic iterator that takes a receiver and a function as arguments. It unwraps the receiver and applies the provided function if no error occurs during unwrapping. The error (if any) is then returned by the Iter function.