Add some functional stuff in go (#263)
* Refactor 'From' interface to 'Converter' struct

The previous 'From' interface has been refactored into a new structure called 'Converter'. This change includes the introduction of a new function type, 'ConverterFunc', and two methods: a constructor method for creating new instances of the Converter struct, and a Convert method that uses the provided ConverterFunc.

* Introduced Runner interface in runner package

A new interface, Runner, has been added to the runner package. This interface integrates the Generator from the gen package and Converter from the from package. It's a generic type that accepts Spec, I, and O as any types.

* Refactoring everything

* Added methods to ConverterFunc and GeneratorFunc types

In this update, we've added a method to each of the ConverterFunc and GeneratorFunc types. These methods allow for direct invocation of these functions with the required context and input parameters. This simplifies their usage by eliminating the need for explicit function calls.

* Refactor runner package

The runner package has been significantly refactored. The context import and the result import have been removed, as well as the entire runner struct and its methods. This simplifies the Runner interface by only including Converter and Generator from 'from' and 'gen' packages respectively.

* Refactored generator function calls in map.go

The generator functions within the MapI and MapO methods have been refactored. The direct function calls have been replaced with the .Gen method to improve code readability and maintainability.