Deleting old stuff (#403)
* Deleting old stuff

* Less cool make but runs more tests

* Updated blacklist and added TypeScript support

The blacklist has been updated with additional directories to ignore. A new feature has been added to support TypeScript files, allowing the user to list TypeScript sources. The code for handling file paths and determining which files should be printed has been refactored into a separate 'printer' struct and associated methods. This improves modularity and makes the code easier to maintain.

* Refactor Makefile for better code organization

The Makefile has been refactored to improve the organization of the code. The changes include replacing a direct 'find' command with a more abstracted function call, and making certain variables optional by using '?='. This should make the file easier to read and maintain.

* Temporarily giving up