Updated build process and file listing (#417)
The Makefile has been updated to use a dynamic list of source files, excluding test files. This change improves the efficiency of the build process by avoiding unnecessary builds.

In addition, the 'devops list' command now supports an '--exclude-tests' flag that allows for exclusion of test files from the output. The handling of Go and TypeScript sources has also been improved to better determine whether or not they should be printed based on their extension and if they are test files.

Lastly, some redundant code in go.mod creation was removed.