Refactoring way too many things (#419)
* WIP refactoring basically everything yet again

* It compiles

* It still compiles

* Move media

* Rename media

* Simplified test descriptions in mediatype tests

Removed the use of fmt.Sprintf for generating test descriptions in the mediatype package. The descriptions are now nil, focusing on testing the functionality rather than creating descriptive strings. This simplifies the code and makes it easier to read and maintain.

* Added new Option type in mediatype package

A new type called 'Option' has been introduced in the mediatype package. This function returns a MediaType from the tdl package. This addition will provide more flexibility and control over media types within the application.

* Basic suite for mediatype matching

* Tests for guessing

* Refactored reader to use options pattern

The reader in the tdl package has been refactored to use an options pattern. This change includes creating a new `readerOptions` struct and modifying the `NewReader` function to accept these options. The media type is now set through this new structure, allowing for more flexibility in its configuration. The `ReadAll` function has also been updated to utilize these changes.

* Refactor reader options and add media type function

Moved the ReaderOption type definition to a more appropriate location in the code. Also, added a new function WithMediaType that returns a MediaType function. This enhances flexibility when working with different media types.