Added unwrapping functionality to plugins
This update introduces the ability to unwrap plugins recursively up to a specified depth. This is achieved through the addition of two new functions, Unwrap and UnwrapAll, which return an array of unwrapped plugins. A constant for setting the maximum unwrap depth has also been added. Additionally, a new test file has been created to ensure that these changes function as expected.