Switched to ProtoReader for command input
The NewReader function in the mediatype package has been replaced with ProtoReader. This change affects how commands are executed in both the cli and generator packages, as well as how tests are run in the reader_test file. The new ProtoReader function is now used to provide input for these commands instead of the previous NewReader function.