Refactor typescript.go to use gen.Cli
The code has been refactored to replace fmt.Stringer with gen.Cli in the supported function. The error message has also been updated to reflect this change. Now, instead of checking if the generator can be inspected, it checks if it's a CLI. The name extraction logic remains unchanged but now operates on an instance of gen.Cli instead of fmt.Stringer.