Added Observable interface and related functions
The new Observable interface has been added to the plugin package. This includes the implementation of Subscribe function in noopObservable struct. Also, Observe, TryObserve, TrySubscribe, and Subscribe functions have been introduced for handling plugins. In addition to this, a minor update was made in progress package by adding TotalObservable interface.