Refactor release caching in GitHub plugin
Significant changes have been made to the release caching function in the GitHub plugin. The code has been simplified by removing unnecessary imports and reducing complexity within the cache() function. This includes replacing a large block of code with a call to cache.ExtractTar(). The variable name 'r' was also changed to 'rel' for clarity.