Added tar extraction functionality and tests (#519)
* Added tar extraction functionality and tests

Introduced a new function to extract tar files. This function reads the tar file, checks if the bin exists, and writes it if not. Accompanying unit tests have been added to ensure correct functionality. Additionally, embedded test data has been included in the cache test suite for more comprehensive testing.

* Refactor release caching in GitHub plugin

Significant changes have been made to the release caching function in the GitHub plugin. The code has been simplified by removing unnecessary imports and reducing complexity within the cache() function. This includes replacing a large block of code with a call to cache.ExtractTar(). The variable name 'r' was also changed to 'rel' for clarity.

* Added file size to cache items

The CacheItem struct now includes a Size field. This change also affects the Get method in the Fs package, which now retrieves and sets the size of the cached file when returning a CacheItem.

* Tweaking stuff

* Kinda fix stuff to a point

* Rework things