Get crd2pulumi working (#533)
* Refactoring bits and bobs

* Copy less unnecessary stuff

* Holy nutsack it works

* Refactor test setup in tool_test.go

The test setup for the "Tool" tests has been refactored. The variable declaration and initialization previously done at the top level of the "Tool" describe block have been moved into a new nested "Execute" describe block. This change ensures that these setup steps are only run for the relevant tests, improving efficiency and readability.

* Updated test function status

Changed the status of a test function from 'Focused' to 'Included'. This allows the test suite to execute all tests instead of just focusing on one.